Splitting Options class into multiple Options###Tab (finished)
