+++ TLM/TLM/UI/MainMenu/DespawnButton.cs
@@ -1,16 +1,22 @@
-﻿using ColossalFramework.UI;
-using TrafficManager.State;
+﻿namespace TrafficManager.UI.MainMenu {
+    using ColossalFramework.UI;
+    using State;
 
-namespace TrafficManager.UI.MainMenu {
     public class DespawnButton : MenuButton {
         public override bool Active => false;
-        public override ButtonFunction Function => Options.disableDespawning ? ButtonFunction.DespawnDisabled : ButtonFunction.DespawnEnabled;
-        public override string Tooltip => Options.disableDespawning ? "Enable_despawning" : "Disable_despawning";
+
+        public override ButtonFunction Function => Options.disableDespawning
+                                                       ? ButtonFunction.DespawnDisabled
+                                                       : ButtonFunction.DespawnEnabled;
+
+        public override string Tooltip =>
+            Options.disableDespawning ? "Enable_despawning" : "Disable_despawning";
+
         public override bool Visible => true;
 
         public override void OnClickInternal(UIMouseEventParameter p) {
             UIBase.GetTrafficManagerTool(true).SetToolMode(ToolMode.None);
-            Options.setDisableDespawning(!Options.disableDespawning);
+            OptionsGameplayTab.SetDisableDespawning(!Options.disableDespawning);
         }
     }
 }
