+++ TLM/TLM/UI/SubTools/SpeedLimitsTool.cs
@@ -416,7 +416,7 @@
             var displayMph = GlobalConfig.Instance.Main.DisplaySpeedLimitsMph;
             displayMph = GUILayout.Toggle(displayMph, Translation.GetString("Display_speed_limits_mph"));
             if (GlobalConfig.Instance.Main.DisplaySpeedLimitsMph != displayMph) {
-                Options.SetDisplayInMph(displayMph);
+                OptionsGeneralTab.SetDisplayInMph(displayMph);
             }
 
             GUILayout.FlexibleSpace();
